backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
pillow==11.2.1
opencv-python==4.11.0.86
sqlalchemy==2.0.41
alembic==1.16.1
aiohttp==3.12.11
python-dotenv==1.1.0
cryptography==45.0.3
pydantic==2.11.5
paddleocr
numpy==2.0.2
paddlepaddle
openai>=1.85.0
requests==2.32.3
matplotlib==3.9.4
fonttools==4.55.3
scikit-image==0.24.0 